Refactored ChemicalPotential files to match class name (from GrandCanonical)
