+++ src/ensemble/ChemicalPotential.h
