+++ src/io/GammaWriter.h
@@ -2,7 +2,7 @@
 #define SRC_IO_GAMMAWRITER_H_
 
 #include "plugins/PluginBase.h"
-#include "ensemble/GrandCanonical.h"
+#include "ensemble/ChemicalPotential.h"
 #include <string>
 #include <fstream>
