+++ src/io/MPI_IOReader.cpp
@@ -17,7 +17,7 @@
 #include "Domain.h"
 #include "ensemble/BoxDomain.h"
 #include "ensemble/EnsembleBase.h"
-#include "ensemble/GrandCanonical.h"
+#include "ensemble/ChemicalPotential.h"
 #include "ensemble/PressureGradient.h"
 #include "Simulation.h"
 #include "molecules/Molecule.h"
