+++ src/io/VISWriter.h
@@ -3,7 +3,7 @@
 
 #include "plugins/PluginBase.h"
 #include "Domain.h"
-#include "ensemble/GrandCanonical.h"
+#include "ensemble/ChemicalPotential.h"
 #include <string>
 #include <list>
