+++ src/io/XyzWriter.h
@@ -3,7 +3,7 @@
 
 #include <string>
 
-#include "ensemble/GrandCanonical.h"
+#include "ensemble/ChemicalPotential.h"
 #include "plugins/PluginBase.h"
 
 /** @brief Writes out an ASCII file in the *.xyz-format containing coordinates of each molecule.
