+++ tools/gui/generators/MDGenerator.cpp
@@ -7,7 +7,7 @@
 
 #include "MDGenerator.h"
 
-#include "ensemble/GrandCanonical.h"
+#include "ensemble/ChemicalPotential.h"
 #include "parallel/DomainDecompBase.h"
 #include "io/CheckpointWriter.h"
 #include "ensemble/PressureGradient.h"
